pyproject.toml
[tool.poetry]
name = "dti-benchmark"
version = "0.2.0"
description = "Benchmark creation for drug-target interaction (DTI) prediction task"
authors = ["Dmitrii Traktirov <ds.traktirov@gmail.com>", "Ellen Kartysheva <>"]

[tool.poetry.dependencies]
dgl = "0.6.*"
hydra-core = ">=1.2"
pandas = "1.3.5"
python = "3.8.*"
numpy = "1.21.6"
rdkit-pypi = "2022.3.2"
scikit-learn = "1.0.2"
torch = "1.11.*"
torchvision = "^0.12.0"
tqdm = "4.64.0"
wandb = "0.12.16"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

